requirements.txt
selenium
requests
webdriver-manager
pytest
allure-pytest